requirements.txt
Django==4.2.16
dj-database-url==2.1.0
whitenoise==6.6.0
Pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2025.1
gunicorn==21.2.0
openai>=1.37.0
python-dotenv>=1.0.1
boto3==1.34.0
django-storages==1.14.2
requests>=2.32.0